requirements.txt
anthropic>=0.40.0
openai>=1.50.0
markdown>=3.7
Pygments>=2.18
pyperclip>=1.9
PyYAML>=6.0
gitpython>=3.1
click>=8.1
